project-1/requirements.txt
matplotlib==3.10.0
numpy==2.2.3
pandas==2.2.3
python-dotenv==1.0.1
scipy==1.15.2
seaborn==0.13.2
